Update Dispatcher.php
Added public function diplomacyRestoreTest to check that you are actually able to restore a realm. This will make sure you're a player, that your realm has inferiors, that at least 1 inferior is dead, and that you are the ruler. Not sure if that last line should be true or false though...