Update House.orm.xml
Don't think I need achievements to actually exist here since it will just be tallying up the character achievements and for that we can just do a few foreach loops with a switch to get all the characters in the house, then get all their achievements, then count them. Dunno if I want to call that every time though, so we may come back to this--may even opt to put this off until later.