2.8.5.0 (#251)
* adding marsh drainage building (proper attempt) (#247)

* Update Economy.php

Marshes are quite useful for agriculture when they are cultivated for it. Though perhaps not the perfect representation of the work gone in to making a marsh for it, irrigation ditches are roughly the same general process for it.

* Update LoadBuildingData.php

Adding in marsh drainage. Higher effort and resources than irrigation ditches, but provides comparable benefits to the fishery.

* Update economy.en.yml

The adding of the marsh drainage's description and such.

* Update Economy.php

Adding in the conditional for marsh drainage.

* Update Economy.php

Fixing tiny error in comment.

* adding marshers building

* Update LoadBuildingData.php

---------

Co-authored-by: jaltonwhitley <60042041+jaltonwhit@users.noreply.github.com>
Co-authored-by: Andrew <Zanaras@users.noreply.github.com>

* re-enabling manual surrender & adding units return/disband on capture (#246)

* Update Dispatcher.php

re-enables manual surrendering.

* Update CharacterManager.php

Sends home/disbands units upon imprisonment, for both manual and in battle captures.

* Update CharacterManager.php

Fix returnUnitHome function call to call it from MilitaryManager

---------

Co-authored-by: Andrew <Zanaras@users.noreply.github.com>

* Update CharacterController.php

Fix edge case for house spawning for characters that have a house but the house doesn't have a home being able to force spawn in a new house.

* Update ActionResolution.php

* Update WarManager.php

* Update WarController.php

* Update Unit.php

* Update BattleGroup.php

* Update CharacterManager.php

* Update ActionResolution.php

* Unit rebasing changes (#250)

* Units can be rebased from their settlements, onto stewarded land, and get supplied by the target.

* Supplier defaults to local settlement.

* Occupied settlements can be rebased to as well.

* some code review and tweaks

* fix unnecessary quote

* getter

* Update CharacterController.php

---------

Co-authored-by: marin (alfatyo) <125774911+alfatyo@users.noreply.github.com>
Co-authored-by: jaltonwhitley <60042041+jaltonwhit@users.noreply.github.com>
Co-authored-by: Rohesie <rohesie@gmail.com>